Add the first functionality for posting drawings
Add the first functionality for posting drawings
- The main page(Главная) should have buttons to go
to the drawing feed, to the form
for posting the drawing, to the profile.
- The share(Поделиться) form should accept the name of the drawing,
a short description of the drawing, a file with the drawing,
a full description of the drawing,
Then you can click on the Добавить рисунок button everything that
you have uploaded/written to the form
should be recorded in the database.
- Display form's(Лента) content should display each of the drawings
that are in the database, but only in the form of a name,
a short description and a button
to go to a detail view of the drawing.
When you click the button, you should go to the detailed view,
where should be displayed the name of picture,
date of publication, brief description and drawing.
- The scoreboard page(Лидеры) should output a stub.
- The profile(Профиль) should output a stub.
- Login(Вход) form must accept a username and password,
then you can click on the button to check the entered data
and log in for further use of the site's functionality.
- Registration(Регистрация) form must accept e-mail,
username, password, repeat password, also the user must click
click Регистрация to register the user
for further login and using the site functionality.
- Exit(Выход) should display a button, when clicked, which
will log out of the account to log in to another account or
for a different purpose.