Add responsive design for low-height screens
Add responsive design for low-height screens.
Add pictures to the feed that fit into the resolution when the screen
is zoomed out.
Add fonts and blocks that fit into the resolution
when the screen is reduced.
Remove the problems associated with moving the control panel
when decreasing the resolution.
Add some cosmetic changes to the control panel and text styles.