Add some styles for view
Add some styles for view.
1. Add a responsive design to the ribbon of pictures. Remove unnecessary
text for minimalism. Add a grid system adaptive for all tape layout
devices. Add animation when hovering over a picture.
2. Add a responsive design for a detailed view of the picture. Add
a grid system for dividing the screen in two. The first contains
the username and picture. The second contains the name of the picture,
a brief description of the picture, full description and date of
publication.