Add a system of profiles, covering the lent with blocks
Add a system of profiles, covering the lent with blocks.
The transition to the user profile can be carried out
in the search of the Лента page, also through the transition by username
when viewing the drawing in detail view, also you can see
your profile by click on button Профиль in navigation bar.
The covering of the tape with blocks now occurs closely,
there are no different spaces and extra indents between pictures.
Add minor design fixes.